feat: create endpoint for home
