+++ src/main/java/com/codeplanks/home360/Home360Application.java
@@ -20,14 +20,4 @@ public class Home360Application {
     public static void main(String[] args) {
         SpringApplication.run(Home360Application.class, args);
     }
-    @GetMapping("/api/v1")
-    public ResponseEntity<Object> apiRoot() {
-        Map<String, Object> body = new LinkedHashMap<>();
-        body.put("message", "Welcome to Home360 version 1 API.");
-        body.put("status", String.valueOf(HttpStatus.OK));
-        body.put("timestamp",
-                new Date().toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime());
-        return new ResponseEntity<>(body, HttpStatus.OK);
-    }
-
 }
